+++ apps/meteor/client/components/CustomScrollbars/VirtuosoScrollbars.tsx
@@ -1,4 +1,4 @@
-import type { ComponentProps, ReactElement, Ref } from 'react';
+import type { ComponentProps, Ref } from 'react';
 import React, { forwardRef } from 'react';
 
 import CustomScrollbars from './CustomScrollbars';
@@ -8,13 +8,9 @@ type VirtuosoScrollbarsProps = ComponentProps<typeof CustomScrollbars>;
 const VirtuosoScrollbars = forwardRef(function VirtuosoScrollbars(
 	{ style, children, ...props }: VirtuosoScrollbarsProps,
 	ref: Ref<HTMLDivElement>,
-): ReactElement {
+) {
 	return (
-		<CustomScrollbars
-			style={{ ...style, flexGrow: 1, overflowY: 'hidden', width: '100%', willChange: 'transform' }}
-			ref={ref}
-			renderView={(viewProps): ReactElement => <div {...viewProps} {...props} />}
-		>
+		<CustomScrollbars style={style} ref={ref} renderView={(viewProps) => <div {...viewProps} {...props} />}>
 			{children}
 		</CustomScrollbars>
 	);