+++ packages/ui-video-conf/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rocket.chat/ui-video-conf
 
+## 7.0.2
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@7.0.2
+  - @rocket.chat/ui-avatar@3.0.2
+  </details>
+
+## 7.0.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@7.0.1
+  - @rocket.chat/ui-avatar@3.0.1
+  </details>
+
 ## 7.0.0
 
 ### Patch Changes